+++ src/main/java/ru/job4j/accidents/repository/AccidentRepository.java
@@ -11,9 +11,9 @@ public interface AccidentRepository {
 
     Collection<Accident> findAll();
 
-    Optional<Accident> findById(int id);
+    Optional<Accident> findById(Long id);
 
-    boolean deleteById(int id);
+    boolean deleteById(Long id);
 
     boolean update(Accident accident);
 }
